requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
groq>=0.4.1
httpx>=0.25.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

